chore: refactor and extract list item
